services/core/requirements.txt
# Core FastAPI and web framework
fastapi==0.111.0
uvicorn[standard]==0.24.0
starlette==0.37.2

# Data validation and serialization
pydantic[email]==2.5.0
python-multipart==0.0.9

# Database support
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# HTTP client
httpx==0.25.2

# Configuration management
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Email validation
email-validator==2.1.0

# Date handling
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# JSON handling
orjson==3.9.10

# Timezone handling
pytz==2023.3

# File operations
aiofiles==23.2.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production server
gunicorn==21.2.0

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Utilities
faker==20.1.0 